Matplotlib tutorial ( beta )
This tutorial contains some mistakes in it. So, please try no to replicate the code thats in it, rather try to understand the code and write it on your own. This way, you're also going to  learn better --  I think personally like doing this and I personally learn better like this.